adjustment + logs to be ignored

+++ openair2/LAYER2/NR_MAC_gNB/main.c
@@ -114,8 +114,8 @@ size_t dump_mac_stats(gNB_MAC_INST *gNB, char *output, size_t strlen, bool reset
   UE_iterator(gNB->UE_info.connected_ue_list, UE) {
     NR_UE_sched_ctrl_t *sched_ctrl = &UE->UE_sched_ctrl;
     NR_mac_stats_t *stats = &UE->mac_stats;
-    const int avg_rsrp = stats->num_rsrp_meas > 0 ? (int)((int64_t)stats->cumul_rsrp / (int64_t)stats->num_rsrp_meas) : 0;
-    const int avg_sinrx10 = stats->num_sinr_meas > 0 ? (int)((int64_t)stats->cumul_sinrx10 / (int64_t)stats->num_sinr_meas) : 0;
+    const int avg_rsrp = stats->num_rsrp_meas > 0 ? stats->cumul_rsrp / (int)stats->num_rsrp_meas : 0;
+    const int avg_sinrx10 = stats->num_sinr_meas > 0 ? stats->cumul_sinrx10 / (int)stats->num_sinr_meas : 0;
 
     output = st_append(output, end, "UE RNTI %04x CU-UE-ID ", UE->rnti);
     if (du_exists_f1_ue_data(UE->rnti)) {
